requirements.txt
torch>=2.2
torchvision>=0.17
timm>=0.9.12
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2
scipy>=1.11.4
SimpleITK>=2.3.1
nibabel>=5.2.0
scikit-image>=0.24.0
torchmetrics>=1.4.0
einops>=0.7.0
tqdm>=4.66.5
pyyaml>=6.0.2
matplotlib>=3.8.4